services/orchestrator_service/requirements.txt
fastapi[standard]>=0.116.1
pydantic>=2.11.7
httpx>=0.28.1
opentelemetry-api>=1.36.0
opentelemetry-sdk>=1.36.0
opentelemetry-instrumentation-fastapi>=0.57b0
opentelemetry-instrumentation-httpx>=0.57b0
opentelemetry-exporter-gcp-trace>=1.9.0
google-cloud-pubsub>=2.31.1
google-cloud-firestore>=2.21.0
